+++ apps/next/components/footer.tsx
@@ -4,32 +4,35 @@ import styles from './footer.module.css';
 import React from 'react';
 import { FaDiscord, FaTwitter } from 'react-icons/fa';
 import { discordLink, instagramLink, twitterLink } from '@app/util/constants';
+import { useRouter } from 'next/router';
 
 export default function Footer() {
+    const router = useRouter();
     return (
         <footer className={styles.footer}>
             <VStack pb="2" spacing="6">
                 <Spacer />
-
-                <VStack spacing="2">
-                    <Heading fontSize="lg">Connect with us! 🤠</Heading>
-                    <Wrap justify="center">
-                        <WrapItem>
-                            <NextLink passHref href="https://twitter.com/PulseBanner?ref_src=twsrc%5Etfw">
-                                <Button as="a" target="_blank" colorScheme="twitter" leftIcon={<FaTwitter />}>
-                                    Follow us on Twitter
-                                </Button>
-                            </NextLink>
-                        </WrapItem>
-                        <WrapItem>
-                            <NextLink passHref href={discordLink}>
-                                <Button as="a" target="_blank" leftIcon={<FaDiscord />}>
-                                    Join our Discord
-                                </Button>
-                            </NextLink>
-                        </WrapItem>
-                    </Wrap>
-                </VStack>
+                {!router.asPath.startsWith('/admin') && (
+                    <VStack spacing="2">
+                        <Heading fontSize="lg">Connect with us! 🤠</Heading>
+                        <Wrap justify="center">
+                            <WrapItem>
+                                <NextLink passHref href="https://twitter.com/PulseBanner?ref_src=twsrc%5Etfw">
+                                    <Button as="a" target="_blank" colorScheme="twitter" leftIcon={<FaTwitter />}>
+                                        Follow us on Twitter
+                                    </Button>
+                                </NextLink>
+                            </WrapItem>
+                            <WrapItem>
+                                <NextLink passHref href={discordLink}>
+                                    <Button as="a" target="_blank" leftIcon={<FaDiscord />}>
+                                        Join our Discord
+                                    </Button>
+                                </NextLink>
+                            </WrapItem>
+                        </Wrap>
+                    </VStack>
+                )}
 
                 <Spacer />
                 <Divider maxW={['56', 'xl']} />