+++ apps/next/features/banner/bannerStreamUp.ts
@@ -1,17 +1,15 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-import NextCors from 'nextjs-cors';
-import { TwitterResponseCode, updateBanner, getBanner, validateTwitterAuthentication } from '@app/util/twitter/twitterHelpers';
-import { flipFeatureEnabled, getBannerEntry, getTwitterInfo } from '@app/util/database/postgresHelpers';
-import { remotionAxios, twitchAxios } from '@app/util/axios';
-import { Prisma } from '@prisma/client';
-import { getAccountsById } from '@app/util/getAccountsById';
-import { TwitchClientAuthService } from '@app/services/TwitchClientAuthService';
-import { AxiosResponse } from 'axios';
-import { env } from 'process';
-import imageToBase64 from 'image-to-base64';
-import { uploadBase64 } from '@app/util/s3/upload';
-// import { download } from '@app/util/s3/download';
-import { checkValidDownload } from '@app/util/s3/validateHelpers';
+import { TwitchClientAuthService } from "@app/services/TwitchClientAuthService";
+import { twitchAxios, remotionAxios } from "@app/util/axios";
+import { getTwitterInfo, flipFeatureEnabled, getBannerEntry } from "@app/util/database/postgresHelpers";
+import { getAccountsById } from "@app/util/getAccountsById";
+import { uploadBase64 } from "@app/util/s3/upload";
+import { checkValidDownload } from "@app/util/s3/validateHelpers";
+import { validateTwitterAuthentication, getBanner, TwitterResponseCode, updateBanner } from "@app/util/twitter/twitterHelpers";
+import { Prisma } from "@prisma/client";
+import { AxiosResponse } from "axios";
+import imageToBase64 from "image-to-base64";
+import { env } from "process";
+import { Feature } from "../Feature";
 import { logger } from '@app/util/logger';
 import { download } from '@app/util/s3/download';
 
@@ -22,18 +20,7 @@ export type TemplateRequestBody = {
     backgroundProps: Record<string, unknown>;
 };
 
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-    // Run the cors middleware
-    // nextjs-cors uses the cors package, so we invite you to check the documentation https://github.com/expressjs/cors
-    await NextCors(req, res, {
-        // Options
-        // methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-        methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-        origin: '*',
-        optionsSuccessStatus: 200, // some legacy browsers (IE11, various SmartTVs) choke on 204
-    });
-
-    const userId: string = req.query.userId as string;
+const bannerStreamUp: Feature<string> = async (userId: string): Promise<string> => {
 
     const accounts = await getAccountsById(userId);
     const twitchUserId = accounts['twitch'].providerAccountId;
@@ -44,13 +31,13 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     const validatedTwitter = await validateTwitterAuthentication(twitterInfo.oauth_token, twitterInfo.oauth_token_secret);
     if (!validatedTwitter) {
         await flipFeatureEnabled(userId, 'banner');
-        return res.status(401).send('Unauthenticated Twitter. Disabling feature banner and requiring re-auth.');
+        return 'Unauthenticated Twitter. Disabling feature banner and requiring re-auth.';
     }
 
     // get the banner info saved in Banner table
     const bannerEntry = await getBannerEntry(userId);
     if (bannerEntry === null || twitterInfo === null) {
-        return res.status(400).send('Could not find banner entry or twitter info for user');
+        return 'Could not find banner entry or twitter info for user';
     }
 
     const authedTwitchAxios = await TwitchClientAuthService.authAxios(twitchAxios);
@@ -75,6 +62,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     // https://stackoverflow.com/a/58158656/10237052
 
+    console.log('bannerUrl', bannerUrl);
+
     // store the current banner in s3
     let dataToUpload: string = bannerUrl === 'empty' ? 'empty' : await imageToBase64(bannerUrl);
 
@@ -89,11 +78,11 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         // check valid download once more
         if (!checkValidDownload(refetch)) {
             // if we are invalid again, fail the request
-            console.log('Corrupt base64 image. Uploading signup image');
+            logger.error('Corrupt base64 image. Uploading signup image');
             const original = await download(env.BANNER_BACKUP_BUCKET, userId);
             if (!checkValidDownload(original)) {
-                console.log('Corrupt signup image. Failing request');
-                return res.status(400).send('Corrupt signup image. Failing request');
+                logger.error('Corrupt signup image. Failing request');
+                return 'Corrupt signup image. Failing request';
             } else {
                 dataToUpload = original;
             }
@@ -106,7 +95,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         await uploadBase64(bucketName, userId, dataToUpload);
     } catch (e) {
         logger.error('Error uploading original banner to S3', { userId });
-        return res.status(500).send('Error uploading original banner to S3.');
+        return 'Error uploading original banner to S3.';
     }
 
     // construct template object
@@ -124,5 +113,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     // post this base64 image to twitter
     const bannerStatus: TwitterResponseCode = await updateBanner(userId, twitterInfo.oauth_token, twitterInfo.oauth_token_secret, base64Image);
-    return bannerStatus === 200 ? res.status(200).send('Set banner to given template') : res.status(400).send('Unable to set banner');
+    return bannerStatus === 200 ? 'Set banner to given template' : 'Unable to set banner';
+
 }
+
+export default bannerStreamUp;