+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.6.0] - 2022-01-22
+
+### Added
+
+- Secured feature endpoints
+- User and webhook admin pages
+
 ## [0.5.6] - 2022-01-22
 
 ### Fixed